+++ CHANGELOG.md
@@ -8,6 +8,12 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+
+- Added support for plugins that send MIDI events back to the host. This plugins
+  such as Cthulhu and Scaler to output notes and CC for another plugin to work
+  with.
+
 ### Changed
 
 - Changed the plugin detection mechanism to support yet another way of